move randomizer_data so that it's not in the build data folder
